# Rukia
Rukia is one of the most powerful military forces of the 21st century. It is also one of the only remaining marxist states by the year 2000.

## Political Structure

Rukia's political structure is made to ensure that no one is at the top with unquestionable power, as everyone is below someone (because the hierarchy is cyclical). The hierarchy goes something like this: The People --> The Parliament --> The Secretary General --> The Council of 10 --> The People.

### Taxation

Taxation in Rukia is arranged in such a way that owning more than 1 Billion (Or 10 million in a decimal currency system) Sei is impossible, as the tax increases in accordance to enforce a hard cap on wealth that is then handed out to the less fortunate.